Refix supporting func types. Fixes #5
We can't use go/printer (I forgot) because we need to inject the package
selector on the front of any indentifier when generating the mocks into
a separate mocks package (the default behavior).

So instead I revert the go/printer usage and add support for func types
into the custom switch.