Use golang.org/x/tools/go/loader instead of go/parser.
It still uses go/parser inside, but has better semantics around loading
packages for which there is no compiled pkg file yet, because it can
load everything from source. It's slower, but more reliable.

Unsure how to add tests for this; we'd need to add a third party import
for which there is source, but have to guarantee that it is unbuilt
(e.g. no pkg/ files exist for it) when the test runs.

Resolves #81