Tell golint to ignore mockery output
The golint commit a5f4a24736 (Ignore generated code) causes it to
ignore files that begin with "Code generated" and end with "DO NOT
EDIT.".   By adding those lines to the prologue, we cause suppress
any warnings golint may flag on the generated code.  This patch also
adds an additional newline between the prologue and the Package
statement, which makes it clear to any other elements of the go
tool-chain that the prologue is not the package comment.

Fixes #155