Adding ability to disable version string in header.
This fixes #334

Various people do not like the fact that all of the mocks in a project
will have diffs when a new version of mockery is installed. This commit
adds the ability to disable adding the version string by setting
`disable-version-string: True` in the yaml file, or by CLI parameter.