Dramatically increase mock generation speed
Mock generation speed is dramatically increased by specifying the
packages to load in config instead of dynamically finding them.

The issue with the previous logic is that a `package.Load` would be
called once per file, even if multiple files were part of the same
package. We can dramatically improve performance by giving a config file
that states which packages should be generated.

Add `with-expecter: True`

Logic fixes