Fix bug when parsing packages config
This fixes #565. There is some bug when a particular package name is applied
to the config section, where viper fails to parse it and the corresponding
attribute in the config struct becomes nil. This is a bug with viper and
I don't really care to find exactly why it happens... I'm guessing some
issue with periods in a key name. https://github.com/spf13/viper/issues/348

The fix is to rely on our direct parsing of the packages variable. An e2e test
has been added to show this has been fixed.

Fix regression with config not existing