Fix issue with self-referential template names
We want template variables to be able to reference other template
variables. This was broken in the current config where we had

```yaml
filename: "{{.MockName}}.go
```

`MockName` is by default `Mock{{.InterfaceName}}`.

We also changed the `structname` variable to be `mockname` to make
it more clear that you can specify the name of the mock directly.

Update docs for `mockname`