Add InterfaceDirRelative template variable
This fixes #623

Also fixed in here was a bug where the template object was being reused. This caused subtle issues where if a templated variable was set to the empty string, the template would simply reuse the last parsed template.

Add docs/fix minor logic flow

Fix bug with test