Include auto-generated files in recursive discovery by default
Fixes #681

There are many legitimate use-cases where mockery needs to include auto-generated files
in its recursive package discovery. We add an `include-auto-generated` parameter that
is set to `true` by default, but can be set to `false` if users want to retain the
old behavior. We are considering the old behavior to be a bug due to the large number
of GitHub issues we'll likely get. The original motivation for adding this behavior
was fixed through a different mechanism, which is why we're considering this to be
a bugfix.