Fix test with config initialization
The test wasn't properly initializing the viper object so the config
wasn't getting decoded into the struct properly.

Also adding more comments to various places to better explain what the
code was doing, as even I, the person who wrote it, was confused as to what
I was doing. How can I expect other people to understand it if I can't?!