Don't recurse into submodules on `recursive: true`
This PR changes mockery to not recurse into paths that contain
a `go.mod` submodule. This is to help support monorepos that contain
many modules. In this case, mockery should not be recursing into these
because it causes issues, like in issue #706.

fix deprecations in mkdocs

remove explicit lint step