Fix issue with file ordering
A bug exists in the logic that determines if `go.mod` exists: if a `*.go` file is returned by the filesystem before `go.mod`, we will not skip the
subtree as intended. To fix this, we always stat for `go.mod` if we find a `*.go` file, as a final check before considering the path to be
a subpackage.