make codecov more permissive
I don't really like coverage tests because it's kind of faulty. Lots of e2e tests aren't registered
in the coverage file, and thus codecov erroneously says we have less coverage than we really do.
I'll make this more permissive so it'll not fail the PR checks, but still have it as a PR check
just as a reference tool.