Add tagging Github action
This commit generally improves CICD in regards to automated tagging.
Instead of manually adding and pushing a tag to trigger a release,
it's controlled through the `mockery-tools.env` config. Furthermore,
doc updates no longer require tags to update the docs page. It too
will key off of the `VERSION` parameter to determine the correct version
to push.