Revert to Go 1.23 in go.mod
It was found in #937 that the upgrade to Go 1.24 in go.mod was both unnecessary
and also a breaking change. I made a dumb mistake in thinking the toolchain
needed to be upgraded to support the generic type alias syntax introduced in
Go 1.24, but all that was needed was upgrading `golang.org/x/tools` which is
the utility that parses the syntax.

This fixes #937.